refactor(view): separate keymap declarations from their actions
ui/view/keymaps.lua carried both jobs: it decided which key binds where, and
it implemented what every key does. At 866 lines, 77% of it was action
bodies, which made the binding rules hard to read and hid a real hazard —
each action closed over tabpage, both buffer numbers and the layout flags,
so a mapping installed for one diff could still be holding a buffer the
session had already replaced.

The action bodies move to ui/view/actions/ unchanged, and take an explicit
context instead of capturing it:

  keymaps.lua   866 -> 214   declarations only
  actions/hunk.lua     252   hunk lookup, patch building, stage/unstage/
                             discard, and the ih textobject
  actions/panes.lua    156   explorer toggle and focus, open-in-previous-tab
  actions/move.lua     147   gm alignment
  actions/diffget.lua  118   do / dp
  actions/stage.lua     70   toggle_stage, toggle_staged_view

Making the capture explicit is the point: the values an action depends on are
now visible in its signature rather than inherited from whichever setup pass
created it.

The other keymap modules are left alone. conflict/keymaps.lua already
delegates to its own action modules, and the explorer and history panels bind
panel-scoped objects that live and die with the panel, so they do not carry
the same hazard.

Behavior is unchanged: every extracted body is identical to its original
modulo the context parameter, and the golden keymap matrix still matches.
